Inlined TranslationPresenter, removed .impl package
