+++ src/test/java/de/vogel612/helper/ui/OverviewModelTest.java
@@ -7,7 +7,6 @@ import static org.mockito.Mockito.*;
 import org.junit.Before;
 import org.junit.Test;
 import de.vogel612.helper.data.Translation;
-import de.vogel612.helper.ui.impl.OverviewModelImpl;
 
 import java.net.URISyntaxException;
 import java.nio.file.Path;
